added Splash Screen, moved call for markers to SplashScreenActivity
